fix unit test error for the status filter on the org homepage
