+++ frontend/app/src/people/widgetViews/__tests__/OrgHeader.spec.tsx
@@ -1,26 +1,67 @@
 import React from 'react';
-import { render, fireEvent, screen } from '@testing-library/react';
+import { render, fireEvent, screen, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { OrgHeader } from 'pages/tickets/org/orgHeader';
+import { OrgBountyHeaderProps } from '../../interfaces.ts';
+import { mainStore } from '../../../store/main.ts';
+
+const MockProps: OrgBountyHeaderProps = {
+  checkboxIdToSelectedMap: {
+    Opened: false,
+    Assigned: false,
+    Paid: false,
+    Completed: false
+  },
+  languageString: '',
+  org_uuid: 'clf6qmo4nncmf23du7ng',
+  onChangeStatus: jest.fn()
+};
 
 describe('OrgHeader Component', () => {
+  beforeEach(() => {
+    jest.spyOn(mainStore, 'getPeopleBounties').mockReset();
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
   it('renders the component correctly', () => {
-    render(<OrgHeader />);
+    render(<OrgHeader {...MockProps} />);
     expect(screen.getByText('Post a Bounty')).toBeInTheDocument();
-    expect(screen.getByLabelText('Status')).toBeInTheDocument();
+    expect(screen.getByText('Status')).toBeInTheDocument();
     expect(screen.getByLabelText('Skill')).toBeInTheDocument();
     expect(screen.getByPlaceholderText('Search')).toBeInTheDocument();
     expect(screen.getByText(/Bounties/i)).toBeInTheDocument();
   });
 
   it('opens the PostModal on "Post a Bounty" button click', async () => {
-    render(<OrgHeader />);
+    render(<OrgHeader {...MockProps} />);
     fireEvent.click(screen.getByText('Post a Bounty'));
   });
 
   it('displays the correct number of bounties', () => {
-    render(<OrgHeader />);
+    render(<OrgHeader {...MockProps} />);
     expect(screen.getByText('284')).toBeInTheDocument();
     expect(screen.getByText('Bounties')).toBeInTheDocument();
+  });
+
+  it('should trigger API call in response to click on status from OrgHeader', async () => {
+    const { getByText } = render(<OrgHeader {...MockProps} />);
+
+    const statusFilter = getByText('Status');
+    expect(statusFilter).toBeInTheDocument();
+
+    fireEvent.click(statusFilter);
+
+    await waitFor(() => {
+      expect(mainStore.getPeopleBounties).toHaveBeenCalledWith({
+        page: 1,
+        resetPage: true,
+        ...MockProps.checkboxIdToSelectedMap,
+        languages: MockProps.languageString,
+        org_uuid: MockProps.org_uuid
+      });
+    });
   });
 });
